Final logout

See merge request brest-gomel-java-training-2020/yankovich_aliaksandr!19

+++ final/Web/model/src/main/java/com/expexchangeservice/model/entities/Review.java
@@ -9,20 +9,17 @@ public class Review {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
-//    @OneToOne(cascade = CascadeType.ALL)
-//    @LazyCollection(LazyCollectionOption.FALSE)
-//    @JoinColumn(name = "idProfile")
+    private long id;
     @Column(name = "username")
     private String username;
     @Column(name = "review")
     private String review;
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
